tests: Move test_near_pm_message_url to test_message_fetch.py.
